Fixes #2
Added method PurgeContent to CacheManager that is called at App Startup.
Modified connection string to add admin mode to execute FlushDB command.

I also added code to handle the disconnect event, but this would only
work in a Single Page App scenario, which is not what we want. So I
commented it out.